backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
paho-mqtt==2.1.0
python-dotenv==1.1.0
python-multipart==0.0.20
aiofiles==24.1.0
cryptography==44.0.2
